Some modifications to include a basic target database and tao.py an interface between chimera and the external scheduler TAO.
Modified model.py to include a new table on the scheduler database,
named Targets. This database stores information on targets that can
later be scheduled and included in other databases. This will
facilitate scheduling operations by chimera. Still need to include 1
more table which will store past information on observations. This will
be usefull for keeping a log of what was done on each night and planing
future observations.

Included tao.py which defines a Chimera class tao.py inside
controlers.scheduler. This Chimera class provides an interface with TAO
which will be usefull for providing a seamless scheduler operation. The
package depends on the python wrapper of skycalc library which can be
easily found in the internet and may be included in a future version of
chimera. The class may operate as a standalone scheduler (in case TAO
is not available) and/or interfacing with TAO.