Improve on scheduler algorithm.
Changed a litle on the philosophy behind making a observational
schedule. The mkqueue task is now responsible for implementing
different observation recipes that can be used to generate an
observation queue. The queue must then be processed by a queue
executor, to perform the observations.

On model.py included mode tables to chimera database, to store more
informations regarding the scheduling algorithms. This includes a
Projects database, which store information regarding the different
observational projects. Other tables where already present on a later
version of the scheduler. Such as Targets database.