setup.py and requirements.txt sanitization.
Well, somehow, someday we started using both setup.py and requirements.txt
for dependencies definition. They were not in sync and apparently we
are not supposed to do this way according to some sources below.

This changes puts all dependencies data into setup.py alone and requirements.txt
now uses a special flag to use the dependencies from setup.py directly.

In addition I removed the Makefile and windows specific setup/NSI as those are
not supported and/or needed.

For Windows we'll have a wheel in the future to install it easier assuming you
already have Python/pip.