Add autoguider action on chimera scheduler controller.
On controller: Add "autoguider" to the configuration list.

On executor: Add actionHandlers

On handlers: Add AutoGuideHandler, a class to deal with the action. If has basically two modes of operation. If the flag "action.stop" is True it will issue an abort signal to the guider. Otherwise it will try to start a guider sequence. It is important to aways "stop" a guiding sequence before slewing the telescope.

On model: Add AutoGuide, a table to store the parameters required for guiding.

On chimera-sched: Update regular expression on function "generateDatabase" so it will read an optional field at the end of the sequence with the guiding parameters between (). Also add a helper for the autoguider.