core: Add new Bus to handle all rpc and pubsub
Bus is a new class that replaces Client/Server and some of Manager's
functionality.

Each Chimera instance have one Bus. All communications passes through
it.

This also removes the Redis transport as it is not supported on Windows.
Now, we use NNG as the only transport. Despite, Bus uses a genetic
Transport class in case we decide to switch transport again.

Serialization is hard coded as JSON for now. The serializers classes
still exists but are not used. Still deciding what to do here.

A lot of changes to improve Ctrl-C handlind were added, in special
adding a mutex based sleep mechanism for ChimeraObject.__main__ threads,
this makes the ssystem snappier on shutdown.