core: pubsub fixes to support generic event subscription
When a Proxy subscribes to an event, it need to know what is the
destination object is at the other side. When you pass a URL that
targets a direct object, like /FakeTelescope/fake, this is simple, but
when you ask for /Telescope/0, Proxy will send a Ping message first and
Pong will return the resolved object if it exists.

Now we are also able to raise an exception when we find that an object
that we want is not available, similar to the Pyro days.

This fixes event handling for generic URLs.